[YS-239] refact: PK 생성 전략을 auto-increment에서 TSID로 변경 (#73)
* chore: add tsid dependency

* feat: add generateTSID() for pk

* refact: replace ulid to tsid in S3PreSignedUrlProvider

* chore: remove primary key generation strategy from entity

* feat: add TSID injection in domain model factory method

* fix: prevent precision loss by writing numbers as strings in JSON responses

* refact: remove configuration for writing numbers as strings

* refact: change ID type from Long to String

* test: alter description

* docs: alter example ID values in Swagger response fields

* refact: abstract TSID generation into an interface to follow DIP

* refact: update existing logic to align with changes

* fix: fix failed test

+++ src/main/kotlin/com/dobby/backend/application/usecase/auth/GenerateTokenWithRefreshTokenUseCase.kt
@@ -20,7 +20,7 @@ class GenerateTokenWithRefreshTokenUseCase(
     )
 
     override fun execute(input: Input): Output {
-        val memberId = tokenGateway.extractMemberIdFromRefreshToken(input.refreshToken).toLong()
+        val memberId = tokenGateway.extractMemberIdFromRefreshToken(input.refreshToken)
         val member = memberGateway.getById(memberId)
         return Output(
             accessToken = tokenGateway.generateAccessToken(member),
